improved installer
can install to any directory now (since program has its own scripts and
colors.txt), and installer installs to chunkymap folder in HOME folder

+++ install-chunkymap-on-ubuntu-from-web.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 cd $HOME
-rm -Rf $HOME/minetest-stuff/minetest-chunkymap
+if [ -d "$HOME/minetest-stuff/minetest-chunkymap" ]; then
+  rm -Rf $HOME/minetest-stuff/minetest-chunkymap
+fi
 CHUNKYMAP_INSTALLER_DIR=$HOME/Downloads/minetest-chunkymap
 if [ ! -d "$HOME/Downloads" ]; then
 	mkdir "$HOME/Downloads"
